requirements.txt
allure-pytest==2.2.1b1
allure-python-commons==2.2.1b1
Appium-Python-Client==0.11
attrs==17.2.0
colorama==0.3.9
custom-text-test-runner==0.1.5
enum34==1.1.6
lxml==4.1.0
namedlist==1.7
pluggy==0.4.0
py==1.4.34
pytest==3.2.1
selenium==3.4.3
selenium-unittest==0.2.3
six==1.10.0
table-printer==0.1.4
xmlrunner==1.7.7
